move display helper into util folder
