+++ src/app/search-result-list/search-result-list.component.ts
@@ -2,7 +2,7 @@ import { Component, ElementRef, OnInit } from '@angular/core';
 import { Router, ActivatedRoute  } from '@angular/router';
 import { AdvancedSearchQuery, SearchResult } from '../search/search.service';
 import { sortByOptions, searchFieldPlaceholders, searchFieldLabels, possibleSearchQueryParams, getFieldOptions, genderOptions } from 'src/app/search-term-values';
-import { eventIcon, filterTitle } from '../display-helpers';
+import { eventIcon, filterTitle } from '../util/display-helpers';
 
 interface SearchQueryParams {
   query?: string,
